Hermes-ErrorFix: fix CI mypy — MSLIE _atr float() on numpy compare-union (numpy 2.x stubs)
Agent: Hermes-ErrorFix
Role: CI mypy repair
Scope: mslie/liquidity_map.py, mslie/swing.py
Why:
 - CI run (actions/runs/32402562963) mypy failed:
   liquidity_map.py:72 / swing.py:80 — Argument 1 to "float" has
   incompatible type "SupportsDunderLT[Any] | SupportsDunderGT[Any]".
   CI installs numpy>=1.26,<3 fresh → numpy 2.x stubs: ndarray[int] indexing
   yields np.floating whose comparisons type as SupportsDunderLT/GT, so
   float(max(np.mean(...), MIN_ATR)) rejected. Local numpy 1.26.4 stubs mask it.
Implementation:
 - _atr() in both files: builtin max() now called on explicit float()
   casts (float(high[i]), float(low[i]), float(close[i-1])) so the compare
   union never reaches float(); tr array dtype=np.float64; float() around
   np.mean result. Version-independent under numpy 1.x AND 2.x stubs.
Verification:
 - mypy src: Success (309 files) locally; probe under numpy2 semantics PASS;
   ruff check/format clean; tests/unit/test_mslie_phase22.py +
   tests/integration/test_mslie_api.py 32/32 pass.
Risk: none (identical runtime math — casts are no-ops on float64 values).
Handoff: none.

+++ src/nexus_scalp/mslie/liquidity_map.py
@@ -66,10 +66,14 @@ def _atr(high: np.ndarray, low: np.ndarray, close: np.ndarray, period: int = 14)
     n = len(high)
     if n < 2:
         return MIN_ATR
-    tr = np.empty(n - 1)
+    tr = np.empty(n - 1, dtype=np.float64)
     for i in range(1, n):
-        tr[i - 1] = max(high[i] - low[i], abs(high[i] - close[i - 1]), abs(low[i] - close[i - 1]))
-    return float(max(np.mean(tr[-period:]), MIN_ATR))
+        tr[i - 1] = max(
+            float(high[i]) - float(low[i]),
+            abs(float(high[i]) - float(close[i - 1])),
+            abs(float(low[i]) - float(close[i - 1])),
+        )
+    return float(max(float(np.mean(tr[-period:])), MIN_ATR))
 
 
 def _clip(v: float, lo: float, hi: float) -> float:
